pyproject.toml
[tool.poetry]
name = "-march-2024"
version = "0.1.0"
description = ""
authors = ["tarastm95 <tarasmazepa95@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.1.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"
pillow = "10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
